pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "remedi",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
extend-select = [
    "E",
    "I",
    "FA",
    "UP",
    "RUF100",
]

[tool.aerich]
tortoise_orm = "app.database.TORTOISE_ORM"
location = "./app/migrations"
src_folder = "./."

[project]
name = "remedi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    "aerich[toml]>=0.8.2",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "pydantic>=2.10.6",
    "pyttings>=2.1.0",
    "tortoise-orm[asyncpg]>=0.24.2",
    "uvicorn>=0.34.0",
]

[project.urls]
Homepage = "https://github.com/ruitcatarino/remedi"
Repository = "https://github.com/ruitcatarino/remedi"
Issues = "https://github.com/ruitcatarino/remedi/issues"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
    "coverage>=7.7.1",
]
